+++ src/BoundedContext/Client/Application/GetClientUseCase.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Src\BoundedContext\Client\Application;
+
+use Src\BoundedContext\Client\Domain\Contracts\ClientRepositoryContract;
+use Src\BoundedContext\Client\Domain\Client;
+use Src\BoundedContext\Client\Domain\ValueObjects\ClientId;
+
+final class GetClientUseCase
+{
+    private $repository;
+
+    public function __construct(ClientRepositoryContract $repository)
+    {
+        $this->repository = $repository;
+    }
+
+    public function __invoke(int $userId): ?Client
+    {
+        $id = new ClientId($userId);
+
+        $user = $this->repository->find($id);
+
+        return $user;
+    }
+}
