+++ src/BoundedContext/Client/Infrastructure/DeleteClientController.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Src\BoundedContext\Client\Infrastructure;
+
+use Illuminate\Http\Request;
+use Src\BoundedContext\Client\Application\DeleteClientUseCase;
+use Src\BoundedContext\Client\Infrastructure\Repositories\EloquentClientRepository;
+
+final class DeleteClientController
+{
+    private $repository;
+
+    public function __construct(EloquentClientRepository $repository)
+    {
+        $this->repository = $repository;
+    }
+
+    public function __invoke(Request $request)
+    {
+        $userId = (int)$request->id;
+
+        $deleteClientUseCase = new DeleteClientUseCase($this->repository);
+        $deleteClientUseCase->__invoke($userId);
+    }
+}
