+++ src/BoundedContext/Client/Infrastructure/GetClientController.php
@@ -1,0 +1,29 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Src\BoundedContext\Client\Infrastructure;
+
+use Illuminate\Http\Request;
+use Src\BoundedContext\Client\Application\GetClientUseCase;
+use Src\BoundedContext\Client\Infrastructure\Repositories\EloquentClientRepository;
+
+final class GetClientController
+{
+    private $repository;
+
+    public function __construct(EloquentClientRepository $repository)
+    {
+        $this->repository = $repository;
+    }
+
+    public function __invoke(Request $request)
+    {
+        $userId = (int)$request->id;
+
+        $getClientUseCase = new GetClientUseCase($this->repository);
+        $user           = $getClientUseCase->__invoke($userId);
+
+        return $user;
+    }
+}
