--- src/BoundedContext/User/Domain/Contracts/UserRepositoryContract.php
@@ -1,24 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Src\BoundedContext\User\Domain\Contracts;
-
-use Src\BoundedContext\User\Domain\User;
-use Src\BoundedContext\User\Domain\ValueObjects\UserEmail;
-use Src\BoundedContext\User\Domain\ValueObjects\UserEmailVerifiedDate;
-use Src\BoundedContext\User\Domain\ValueObjects\UserId;
-use Src\BoundedContext\User\Domain\ValueObjects\UserName;
-
-interface UserRepositoryContract
-{
-    public function find(UserId $id): ?User;
-
-    public function findByCriteria(UserName $userName, UserEmail $userEmail): ?User;
-
-    public function save(User $user): void;
-
-    public function update(UserId $userId, User $user): void;
-
-    public function delete(UserId $id): void;
-}
