--- src/BoundedContext/User/Domain/ValueObjects/UserRememberToken.php
@@ -1,20 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Src\BoundedContext\User\Domain\ValueObjects;
-
-final class UserRememberToken
-{
-    private $value;
-
-    public function __construct(?string $rememberToken)
-    {
-        $this->value = $rememberToken;
-    }
-
-    public function value(): ?string
-    {
-        return $this->value;
-    }
-}
